Update NAT Router, now working.
This adds support for
- ICMP, not tested
- a next hop, tested and working, needed for next things
- ICMP echo, tested and working
- FTP(passiv and active), tested and working

Note for the next hop:
The NAT part reset the DST_ANNO. For the intern -> extern path, the routing part
set it and it is needed by the ARPQuerier. If there is a next hop the routing
part set in the DST_ANNO the IP address of the next hop and the ARPQuerier set
the right dst. MAC address.
So i had to disable it.
For the for the other way around the routing part needs the updated DST_ANNO.
It is now set by GetIPAddress.

veth3 is now in the private network and veth1 is in the public network.
There just switched.